+++ src/CategoryLoopException.php
@@ -6,12 +6,17 @@ use Mediawiki\DataModel\Pages;
 
 /**
  * Class CategoryLoopException
+ *
+ * @access private
+ *
+ * @since 0.7
+ *
  * @package Mediawiki\Api
  */
 class CategoryLoopException extends \Exception {
 
 	/** @var Pages */
-	protected $categoryPath;
+	private $categoryPath;
 
 	/**
 	 * @param Pages $path
@@ -23,6 +28,9 @@ class CategoryLoopException extends \Exception {
 	/**
 	 * Get the path of Pages that comprise the category loop. The first item in this list is also a
 	 * child page of the last item.
+	 *
+	 * @since 0.7
+	 *
 	 * @return Pages The set of category Pages that comprise the category loop.
 	 */
 	public function getCategoryPath() {