+++ tests/integration/UserIntegrationTest.php
@@ -16,12 +16,14 @@ class UserIntegrationTest extends PHPUnit_Framework_TestCase {
 	private static $localApiUser;
 
 	public static function setUpBeforeClass() {
+
 		parent::setUpBeforeClass();
 		$strTime = strval( time() );
 		self::$localApiUser = new ApiUser( 'TestUser - ' . strval( time() ), $strTime . '-pass' );
 	}
 
 	public function testCreateUser() {
+
 		$factory = TestEnvironment::newDefault()->getFactory();
 		$createResult = $factory->newUserCreator()->create(
 			self::$localApiUser->getUsername(),